feat: change entities from functions to const
