+++ services/core/constants.ts
@@ -2,3 +2,4 @@ export const OPENAI_API_KEY_PARAM_NAME = "GLOBAL_OPENAI_API_KEY";
 export const LANGCHAIN_API_KEY_PARAM_NAME = "LANGCHAIN_API_KEY";
 export const GITHUB_SECRET_PARAM_NAME = "GITHUB_WEBHOOK_SECRET";
 export const GITHUB_SIGNATURE_HEADER_KEY = "x-hub-signature-256";
+export const TABLE_NAME = "crgpt-data";
