+++ services/core/entities/ReviewDemoCounterEntity.ts
@@ -1,0 +1,20 @@
+import { Entity } from "dynamodb-toolbox";
+import { Table } from "./Table";
+
+export const ReviewDemoCounterEntity = new Entity({
+  name: "demoReviewDayCounterEntity",
+  attributes: {
+    PK: { partitionKey: true, hidden: true, prefix: "DATE#" },
+    SK: { sortKey: true, hidden: true, default: "DEMO" },
+    date: [
+      "PK",
+      0,
+      {
+        type: "string",
+        default: () => new Date().toDateString(),
+      },
+    ],
+    dailyCount: { type: "number" },
+  },
+  table: Table,
+} as const);
