+++ services/core/entities/index.ts
@@ -1,2 +1,3 @@
-export { getUserEntity } from "./userEntity";
-export { getDemoReviewDayCounterEntity } from "./demoReviewDayCounterEntity";
+export { UserEntity } from "./UserEntity";
+export { ReviewDemoCounterEntity } from "./ReviewDemoCounterEntity";
+export { ReviewEntity } from "./ReviewEntity";
